# docs/swagger.yaml
basePath: /
definitions:
...
  abude-backend_internal_pkg_inventories_category.Category:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.Category'
      updatedAt:
        type: string
    type: object
  abude-backend_internal_pkg_inventories_category.CategoryDTO:
    properties:
      description:
        type: string
      name:
        type: string
      parent:
        type: integer
    required:
    - name
    type: object
...
  common.GeneralResponse:
    properties:
      message:
        type: string
      result: {}
    type: object
  common.PaginatedResponse:
    properties:
      metadata:
        $ref: '#/definitions/pagination.Metadata'
      result:
        items: {}
        type: array
    type: object
  company.Company:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      region:
        type: string
      updatedAt:
        type: string
    type: object
...
  handover.Handover:
    properties:
      cashReceived:
        type: integer
      cashReturned:
        type: integer
      createdAt:
        type: string
      date:
        type: string
      expensesTotal:
        type: integer
      id:
        type: integer
      note:
        type: string
      outlet:
        $ref: '#/definitions/outlet.Outlet'
      purchasesTotal:
        type: integer
      salesTotal:
        type: integer
      shift:
        $ref: '#/definitions/shift.Shift'
      updatedAt:
        type: string
    type: object
  handover.HandoverDTO:
    properties:
      cashReceived:
        type: integer
      cashReturned:
        type: integer
      date:
        format: date-time
        type: string
      note:
        type: string
      outlet:
        type: integer
      shift:
        type: integer
    required:
    - cashReceived
    - date
    - note
    - outlet
    - shift
    type: object
...
  outlet.Outlet:
    properties:
      address:
        type: string
      company:
        $ref: '#/definitions/company.Company'
      createdAt:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  pagination.Metadata:
    properties:
      count:
        type: integer
      hasNext:
        type: boolean
      hasPrev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  product.Ingredient:
    properties:
      base:
        $ref: '#/definitions/product.Product'
      createdAt:
        type: string
      id:
        type: integer
      ingredient:
        $ref: '#/definitions/product.Product'
      quantity:
        type: number
      updatedAt:
        type: string
    type: object
  product.IngredientDTO:
    properties:
      product:
        type: integer
      quantity:
        minimum: 0
        type: number
    required:
    - product
    - quantity
    type: object
  product.Product:
    properties:
      category:
        $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.Category'
      company:
        $ref: '#/definitions/company.Company'
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      ingredients:
        items:
          $ref: '#/definitions/product.Ingredient'
        type: array
      isDefault:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        type: boolean
      type:
        enum:
        - purchase
        - sale
        type: string
      unit:
        type: string
      updatedAt:
        type: string
    type: object
  product.ProductDTO:
    properties:
      category:
        type: integer
      company:
        type: integer
      description:
        type: string
      ingredients:
        items:
          $ref: '#/definitions/product.IngredientDTO'
        type: array
      isDefault:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        type: boolean
      type:
        enum:
        - purchase
        - sale
        type: string
      unit:
        type: string
    required:
    - company
    - ingredients
    - name
    - price
    - stock
    - type
    - unit
    type: object
...
  shift.Shift:
    properties:
      createdAt:
        type: string
      description:
        type: string
      endTime:
        type: string
      id:
        type: integer
      name:
        type: string
      startTime:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact:
    email: infamous0192@gmail.com
    name: Dwa Meizadewa
  description: Aplikasi Backend (API) untuk project Abude
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Abude Backend
  version: 2.0.0
paths:
...
  /api/category:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: parent
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.Category'
                  type: array
              type: object
      security:
      - JWT: []
      summary: Get All Category
      tags:
      - Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.CategoryDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.Category'
              type: object
      security:
      - JWT: []
      summary: Create Category
      tags:
      - Products
  /api/category/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.Category'
              type: object
      security:
      - JWT: []
      summary: Delete Category
      tags:
      - Products
    get:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.Category'
      security:
      - JWT: []
      summary: Get One Category
      tags:
      - Products
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.CategoryDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/abude-backend_internal_pkg_inventories_category.Category'
              type: object
      security:
      - JWT: []
      summary: Update Category
      tags:
      - Products
...
  /api/handover:
    get:
      consumes:
      - application/json
      parameters:
      - format: date-time
        in: query
        name: endDate
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: outlet
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: shift
        type: integer
      - format: date-time
        in: query
        name: startDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/handover.Handover'
                  type: array
              type: object
      security:
      - JWT: []
      summary: Get All Handover
      tags:
      - Handovers
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handover.HandoverDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/handover.Handover'
              type: object
      security:
      - JWT: []
      summary: Create Handover
      tags:
      - Handovers
  /api/handover/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Handover ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/handover.Handover'
              type: object
      security:
      - JWT: []
      summary: Delete Handover
      tags:
      - Handovers
    get:
      consumes:
      - application/json
      parameters:
      - description: Handover ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handover.Handover'
      security:
      - JWT: []
      summary: Get One Handover
      tags:
      - Handovers
    put:
      consumes:
      - application/json
      parameters:
      - description: Handover ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handover.HandoverDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/handover.Handover'
              type: object
      security:
      - JWT: []
      summary: Update Handover
      tags:
      - Handovers
...
  /api/product:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: category
        type: integer
      - in: query
        name: company
        type: integer
      - in: query
        name: default
        type: boolean
      - in: query
        name: keyword
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - enum:
        - purchase
        - sale
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/product.Product'
                  type: array
              type: object
      security:
      - JWT: []
      summary: Get All Product
      tags:
      - Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.ProductDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/product.Product'
              type: object
      security:
      - JWT: []
      summary: Create Product
      tags:
      - Products
  /api/product/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/product.Product'
              type: object
      security:
      - JWT: []
      summary: Delete Product
      tags:
      - Products
    get:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
      security:
      - JWT: []
      summary: Get One Product
      tags:
      - Products
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.ProductDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/product.Product'
              type: object
      security:
      - JWT: []
      summary: Update Product
      tags:
      - Products
...
schemes:
- https
- http
securityDefinitions:
  JWT:
    description: Authorization for JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"